+++ example/lib/pages/welcome_page.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_motionly/common/utils.dart';
 import 'package:flutter_motionly/widget/button/focus_button.dart';
+import 'package:flutter_motionly/widget/loaders/square_matrix_loader.dart';
 import 'package:flutter_motionly/widget/text/animated_text.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:url_launcher/url_launcher.dart';